✨ refactor(SessionDetailPage): streamline schema handling and introduce SchemaProvider
- Removed direct schema fetching and state management from SessionDetailPage.
- Introduced SchemaProvider to manage schema state and loading logic.
- Created SessionDetailContent component to encapsulate chat and artifact display.
- Updated Artifact component to utilize schema from SchemaProvider.
- Improved loading and error handling states for better user experience.
