+++ frontend/apps/app/components/Chat/services/messageServiceClient.ts
@@ -20,7 +20,7 @@ const realtimeMessageSchema = v.object({
   id: v.string(),
   design_session_id: v.pipe(v.string(), v.uuid()),
   content: v.string(),
-  role: v.picklist(['user', 'assistant', 'schema_version']),
+  role: v.picklist(['user', 'assistant', 'schema_version', 'error']),
   user_id: v.nullable(v.string()),
   created_at: v.string(),
   updated_at: v.string(),
