+++ frontend/apps/app/components/SessionDetailPage/components/SessionDetailContent/index.ts
@@ -1,0 +1,1 @@
+export { SessionDetailContent } from './SessionDetailContent'
